CORE-1190 many-many dma bug fixes

+++ driver/dma.h
@@ -133,7 +133,7 @@ struct altr_dma_dev {
     void __iomem* msgdma_base;
     void __iomem* prefetcher_base;
     struct pci_dev* pci_device;
-    unsigned long max_transfer_size;
+    unsigned int max_transfer_size;
 
     // This covers transfers_ready_for_hardware, transfers_on_hardware and post_hardware
     spinlock_t hardware_lock;
